Changed comment
I had `drawAttention` in the window update because I was trying to get the port listener to work but I forgot to change it back to `focused` when committing. Doesn't affect functionality right now but later down the road if this does get added `focused` is the update param needed.